api/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
requests==2.31.0
pydantic==1.10.9
python-dotenv==1.0.0